project-2/part2/requirements.txt
Django==4.2
djangorestframework==3.14.0
requests==2.28.2
redis==4.5.4
django_redis==5.2.0
rdbtools==0.1.15
